Move plot_pulls*.py to scripts
